content_service/requirements.txt
pydantic==2.7.1
pydantic-settings==2.3.4
fastapi==0.111.0
elasticsearch[async]==8.13.2
elasticsearch-dsl[async]==8.16.0
redis==5.0.4
uvicorn==0.30.6
fastapi-cache2==0.2.2
backoff==2.2.1
python-jose